(/usr/share/games/fortunes/linux) % #if _FP_W_TYPE_SIZE < 32 #error "Here's a nickel kid.  Go buy yourself a real computer." #endif 	-- linux/arch/sparc64/double.h
