+++ package.json
@@ -56,6 +56,7 @@
 		"mathjs": "^13.1.1",
 		"node-sass": "^9.0.0",
 		"phaser": "^3.88.2",
+		"phaser3-rex-plugins": "^1.80.14",
 		"sass": "^1.79.1",
 		"svelte-persisted-store": "^0.11.0",
 		"xstate": "^5.18.1"
