+++ src/game/scenes/Boot.ts
@@ -9,7 +9,11 @@ export class Boot extends Scene {
 		//  The Boot Scene is typically used to load in any assets you require for your Preloader, such as a game logo or background.
 		//  The smaller the file size of the assets, the better, as the Boot Scene itself has no preloader.
 
-		//this.load.image('background', 'assets/bg.png');
+		scene.load.plugin(
+			'rexstatemanagerplugin',
+			'https://raw.githubusercontent.com/rexrainbow/phaser3-rex-notes/master/dist/rexstatemanagerplugin.min.js',
+			true
+		);
 		// initialize the random map generator first
 		let initialGameState = engine.igs();
 		// draw the map.
