+++ src/game/scenes/Game.ts
@@ -1,6 +1,8 @@
 import { EventBus } from '../EventBus';
 import { Scene } from 'phaser';
 import { drawCity } from '../engine/Canvas';
+import { user, gameStatePersisted } from '../stores.ts';
+
 export class Game extends Scene {
 	camera: Phaser.Cameras.Scene2D.Camera;
 	background: Phaser.GameObjects.Image;
@@ -11,6 +13,40 @@ export class Game extends Scene {
 	}
 
 	create() {
+		let gameStateTemp = {
+			persistedState: gameStatePersisted.state,
+			state: 'waiting',
+			time: 0,
+			worldmap: { cities: [{ id: 0, name: 'anonymous', tier: 5, difficulty: 'rustic' }] },
+			user: {
+				name: gameStatePersisted.user_name,
+				id: gameStatePersisted.user_id,
+				health: 100,
+				currentLocation: { city: 0, loc: { name: 'home', pos: [] } },
+				energy: {
+					social: 100,
+					asocial: 100,
+					weird: 100
+				},
+				neuro: gameStatePersisted.user_neuro,
+				social: gameStatePersisted.user_social,
+				asocial: gameStatePersisted.user_asocial,
+				weird: gameStatePersisted.user_weird
+			},
+
+			allUsers: new Array(),
+			locationUserMap: {
+				0: {
+					university: new Array(),
+					home: new Array(),
+					library: new Array(),
+					suicide_park: new Array(),
+					dance: new Array()
+				}
+			}
+		};
+		this.statemgr = new StateManager(config);
+
 		this.camera = this.cameras.main;
 		this.camera.setBackgroundColor(0x00ff00);
