+++ src/game/scenes/MainMenu.ts
@@ -14,6 +14,7 @@ export class MainMenu extends Scene {
 	}
 
 	create() {
+		var states = new StateManager(config);
 		this.background = this.add.image(512, 384, 'background');
 
 		this.logo = this.add.image(512, 300, 'logo').setDepth(100);
